+++ app/src/main/java/ch/epfl/reminday/security/PromptUserUnlockImpl.kt
@@ -1,4 +1,4 @@
-package ch.epfl.reminday.data.security
+package ch.epfl.reminday.security
 
 import android.content.Context
 import android.util.Log
